WebSockets/tests/submissions/Opera/interfaces/CloseEvent/004.html

<!doctype html>
<title>WebSockets: wasClean, true</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../constants.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/ws/echo');
  ws.onopen = function(e) {
    ws.send('Test');
  }
  ws.onmessage = function(e) {
    ws.close();
  }
  ws.onclose = function(e) {
    assert_equals(e.wasClean,true);
    end();
  }
});
</script>
